modularizacion de ServerController.dart en ServerAppsController.dart, ServerCommandController.dart, ServerConnectionController.dart y  ServerTemplateController.dart
