+++ pubspec.yaml
@@ -45,6 +45,7 @@ dependencies:
   go_router: ^17.2.0
   flutter_secure_storage: ^9.2.2
   file_picker: ^8.0.0
+  get_it: ^9.2.1
 
 dev_dependencies:
   flutter_test:
